requirements.txt
pandas
numpy
matplotlib
scikit-learn
scipy
scikit-surprise
joblib
streamlit